+++ source/constants/emojis/services.ts
@@ -1,7 +1,0 @@
-export default Object.freeze({
-	notices: {
-		welcome: {
-			understood: "✅",
-		},
-	},
-} as const);