+++ source/constants/media.ts
@@ -1,0 +1,13 @@
+export default Object.freeze({
+	images: {
+		// TODO(vxern): Change this before deploying.
+		//banner: "https://raw.githubusercontent.com/LearnRomanian/rost/refs/heads/main/assets/images/Discord%20(Banner)%20%E2%80%94%C2%A0Small.png",
+		banner: "https://media.discordapp.net/attachments/1398666723111997512/1399849870885519473/Discord_Banner_Small.png?ex=688a7f49&is=68892dc9&hm=245755e4b969f025631a95589a05ca080a7141c8add5cf4b8f7abc80c3b4649a&=&format=webp&quality=lossless&width=2000&height=500",
+	},
+	gifs: {
+		done: "https://media.tenor.com/3wJ_QISYvxEAAAAC/done-all.gif",
+		followRules: "https://media.tenor.com/mPxD454ittYAAAAC/rules.gif",
+		chaosWithoutRules: "https://media.tenor.com/_sFydMmz7YgAAAAd/loop-rules-rule.gif",
+		welcome: "https://media.tenor.com/u4Pg5kqdaIAAAAAC/welcome.gif",
+	},
+} as const);