+++ source/library/stores/services.ts
@@ -9,10 +9,8 @@ import { EntryService } from "rost/services/entry";
 import { InteractionRepetitionService } from "rost/services/interaction-repetition";
 import { LavalinkService } from "rost/services/lavalink";
 import { MusicService } from "rost/services/music";
+import { EntryNoticeService } from "rost/services/notices/entry";
 import { InformationNoticeService } from "rost/services/notices/information";
-import { ResourceNoticeService } from "rost/services/notices/resources";
-import { RoleNoticeService } from "rost/services/notices/roles";
-import { WelcomeNoticeService } from "rost/services/notices/welcome";
 import { ReportPromptService } from "rost/services/prompts/reports";
 import { ResourcePromptService } from "rost/services/prompts/resources";
 import { SuggestionPromptService } from "rost/services/prompts/suggestions";
@@ -35,9 +33,7 @@ interface LocalServices {
 	readonly entry: EntryService;
 	readonly music: MusicService;
 	readonly informationNotices: InformationNoticeService;
-	readonly resourceNotices: ResourceNoticeService;
-	readonly roleNotices: RoleNoticeService;
-	readonly welcomeNotices: WelcomeNoticeService;
+	readonly entryNotices: EntryNoticeService;
 	readonly reportPrompts: ReportPromptService;
 	readonly resourcePrompts: ResourcePromptService;
 	readonly suggestionPrompts: SuggestionPromptService;
@@ -92,9 +88,7 @@ class ServiceStore {
 			entry: new Map(),
 			music: new Map(),
 			informationNotices: new Map(),
-			resourceNotices: new Map(),
-			roleNotices: new Map(),
-			welcomeNotices: new Map(),
+			entryNotices: new Map(),
 			reportPrompts: new Map(),
 			resourcePrompts: new Map(),
 			suggestionPrompts: new Map(),
@@ -165,25 +159,11 @@ class ServiceStore {
 			this.#local.informationNotices.set(guildId, service);
 		}
 
-		if (guildDocument.hasEnabled("resourceNotices")) {
-			const service = new ResourceNoticeService(this.#client, { guildId });
+		if (guildDocument.hasEnabled("entryNotices")) {
+			const service = new EntryNoticeService(this.#client, { guildId });
 			services.push(service);
 
-			this.#local.resourceNotices.set(guildId, service);
-		}
-
-		if (guildDocument.hasEnabled("roleNotices")) {
-			const service = new RoleNoticeService(this.#client, { guildId });
-			services.push(service);
-
-			this.#local.roleNotices.set(guildId, service);
-		}
-
-		if (guildDocument.hasEnabled("welcomeNotices")) {
-			const service = new WelcomeNoticeService(this.#client, { guildId });
-			services.push(service);
-
-			this.#local.welcomeNotices.set(guildId, service);
+			this.#local.entryNotices.set(guildId, service);
 		}
 
 		if (guildDocument.hasEnabled("alerts")) {